+++ sdk/src/wallet/account/types/mod.rs
@@ -15,7 +15,7 @@ use crypto::keys::slip10::Chain;
 use serde::{Deserialize, Deserializer, Serialize};
 
 pub use self::{
-    address::{AccountAddress, AddressWithUnspentOutputs, AddressWrapper},
+    address::{AccountAddress, AddressWithUnspentOutputs},
     balance::{
         AccountBalance, AccountBalanceDto, BaseCoinBalance, BaseCoinBalanceDto, NativeTokensBalance,
         NativeTokensBalanceDto, RequiredStorageDeposit,
