+++ sdk/src/wallet/events/types.rs
@@ -8,9 +8,12 @@ use crate::{
     client::api::PreparedTransactionDataDto,
     types::{
         api::core::response::OutputWithMetadataResponse,
-        block::payload::transaction::{dto::TransactionPayloadDto, TransactionId},
+        block::{
+            address::AddressWrapper,
+            payload::transaction::{dto::TransactionPayloadDto, TransactionId},
+        },
     },
-    wallet::account::types::{address::AddressWrapper, InclusionState, OutputDataDto},
+    wallet::account::types::{InclusionState, OutputDataDto},
 };
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
 pub struct Event {
